Fixing typos and making it work for single GPU training.

+++ first.py
@@ -18,7 +18,7 @@ def data_gen(V, batch, nbatches):
     Generate random data for a src-tgt copy task.
     """
     for i in range(nbatches):
-        data = torch.from_numpy(numpy.random.randint(1, V, size=(batch, 10)))
+        data = torch.from_numpy(numpy.random.randint(1, V, size=(batch, 10))).long()
         data[:, 0] = 1
         src = Variable(data, requires_grad=False)
         tgt = Variable(data, requires_grad=False)
